+++ R/mvtb_summary.R
@@ -40,8 +40,8 @@ gbm.ri <- function(object,n.trees=NULL,relative="col",...){
 #' @param relative If \code{"col"}, each column sums to 100. If \code{"tot"}, the whole matrix sums to 100 (a percent). Otherwise, the raw reductions in SSE are returned.
 #' @return Matrix of (relative) influences.
 #' @details 
-#' (Relative) influences are the usual reductions in SSE due to splitting on each predictor. 
-#' If \code{"weighted"=TRUE}, influences are the raw influences weighted by the covariance explained in all pairs of outcomes by that predictor. 
+#' 'The 'influence' of a predictor is the reduction in SSE due to splitting on each predictor, summed over all trees.
+#' If \code{"weighted"=TRUE}, the influence is weighted by the covariance explained in all pairs of outcomes by that predictor. 
 #' This allows predictor selection to be informed by the covariance explained. 
 #' Different weighting types are possible, see \code{?mvtb}.
 #' @export 
@@ -135,12 +135,12 @@ summary.mvtb <- function(object,print=TRUE,n.trees=NULL,relative="col",covex=TRU
 #' If predictors are not independent, the decomposition of covariance explained is only approximate (like the decomposition of R^2 by each predictor in a linear model). 
 #' If interaction.depth > 1, the following heuristic is used: the covariance explained by the tree is assigned to the predictor with the largest influence in each tree.
 #'
-#' Note that different distances measures (e.g. \code{"manhattan"}, \code{"euclidean"}) provide different ways to measure distances between 
+#' Note that different distances measures (e.g. \code{"manhattan"}, \code{"euclidean"}) provide different ways to measure (dis)similarities between 
 #' the covariance explained patterns for each predictor. See \code{?dist} for further details.
 #' After the distances have been computed, \code{hclust} is used to form clusters. 
 #' Different clustering methods (e.g. \code{"ward.D"}, \code{"complete"}) generally group rows and columns differently (see \code{?hclust} for further details).
 #' It is suggested to try different distance measures and clustering methods to obtain the most interpretable solution. 
-#' The defaults are for 'manhattan' distances and \code{"ward.D2"} clustering, which seem to provide reasonable solutions in many cases.
+#' The defaults are for \code{"euclidean"} distances and \code{"complete"} clustering.
 #' Transposing the rows and columns may also lead to different results.
 #'
 #' A simple heatmap of the clustered matrix can be obtained by setting \code{plot=TRUE}. Details of the plotting procedure are available via \code{mvtb.heat}.